Changes to improve the accessibility of the "Notes for translators:" control.
* Added the AutoWrappingTextCtrl class. It is based on the
AutoWrappingText class but is derived from the AutoWrappingText class
but is derived from the wxTextCtrl class instead of the wxStaticText
class. It creates a read only edit control.

* Modified the ExtractedCommentSidebarBlock class so that it uses the
AutoWrappingTextCtrl class instead of the AutoWrappingText class for the
m_comment member variable.

These changes improve the accessibility of the "Notes for translators:"
control and fix Freedom Scientific bug 86987 (POedit tools requires a
improvement in navigation to easily read Translators Notes).