Changes made by the refresh-pot.sh script.
In commit 9d22b040de8e90e99647ccf2dd155c86e41c3160 I applied the same
changes I made to improve the accessibility of the Notes for translators
control to the Comment control. As part of this change I added the Go to
comment menu. I ran the refresh-pot.sh script once again to make the
necessary updated to the PO files. Note that they new entries are not
yet localized.